all: reorganize code a bit
* Move BaseMiddleware to middleware.go
* Merge util_http_helpers into reverse_proxy (the file got small, and
  it's mostly used and fitting there)
* Merge redis_notifier_outbound into redis_signals, now the file holds
  both outgoing and incoming redis notifications

Updates #334.
