+++ host_checker.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/jeffail/tunny"
-	"github.com/pmylund/go-cache"
+	cache "github.com/pmylund/go-cache"
 )
 
 const (
