--- util_http_helpers_test.go
@@ -1,60 +1,0 @@
-package main
-
-import (
-	"net/http"
-	"testing"
-)
-
-func TestRequestIP(t *testing.T) {
-	tests := []struct {
-		remote, real, forwarded, want string
-	}{
-		// missing ip or port
-		{want: ""},
-		{remote: ":80", want: ""},
-		{remote: "1.2.3.4", want: ""},
-		{remote: "[::1]", want: ""},
-		// no headers
-		{remote: "1.2.3.4:80", want: "1.2.3.4"},
-		{remote: "[::1]:80", want: "::1"},
-		// real-ip
-		{
-			remote: "1.2.3.4:80",
-			real:   "5.6.7.8",
-			want:   "5.6.7.8",
-		},
-		{
-			remote: "[::1]:80",
-			real:   "::2",
-			want:   "::2",
-		},
-		// forwarded-for
-		{
-			remote:    "1.2.3.4:80",
-			forwarded: "5.6.7.8, px1, px2",
-			want:      "5.6.7.8",
-		},
-		{
-			remote:    "[::1]:80",
-			forwarded: "::2",
-			want:      "::2",
-		},
-		// both real-ip and forwarded-for
-		{
-			remote:    "1.2.3.4:80",
-			real:      "5.6.7.8",
-			forwarded: "4.3.2.1, px1, px2",
-			want:      "5.6.7.8",
-		},
-	}
-	for _, tc := range tests {
-		r := &http.Request{RemoteAddr: tc.remote, Header: http.Header{}}
-		r.Header.Set("x-real-ip", tc.real)
-		r.Header.Set("x-forwarded-for", tc.forwarded)
-		got := requestIP(r)
-		if got != tc.want {
-			t.Errorf("requestIP({%q, %q, %q}) got %q, want %q",
-				tc.remote, tc.real, tc.forwarded, got, tc.want)
-		}
-	}
-}
